requirements.txt
flask
flask_sqlalchemy
requests
python-dotenv
flask_script
flask_migrate
psycopg2
gunicorn